Fixing randomization to handle previously done workouts.
Motivation:
When a random workout is choses we should also check that it was not
performed during the last workout session to get variation in our
workouts.

Modifications:
Adding a check to see if the current UserWorkout contains the workout
picked randomly. If it has done in the last workout, then remove the
NSManagedObjectID from the available workouts to chooose from.