Fixing async timer tests.
Motivation:
I had copied a test class as they were very similar. One was a countdown
timer and the other a normal timer. The test methods them selves were
named the same. When adding a clear debug statement I noticed that the
first test was getting invoked twice when I expected the second test to
be run.

Modifications:
I had a closure parameter named 'timer' and also a local constant
named 'timer'. This seems to have assigned and kept the first Timer
parameter instance in memory. It was this closure that was getting used
by the second test. By simply using a different name for the parameter
things work as expected.

Result:
All tests pass.