Fixing interval duration time when added via UI
Motivation:
When adding an interval workout using the app (not preloaded from json)
the duration for the workout and the rest workout were taken from the
selected workouts. But changing these existing workouts would also
affect those workout's duration which is not what we want. Instead we
should be creating a new workout for the work/rest workouts in a
interval workout.