Updated code to Swift 2.0
Motivation:
A major update to to Swift was done and the code base should be updated
to reflect this.

Modifications:
Ran the default conversion tool that is provided with Xcode and also
make manual changes that were not done as part of the default
conversion.

There is still more clean up do be done and to verify that everything
works well. I've not yet on iOS 9.1 but will update shortly to verify
that this also works.

Result:
The test complete successfully:
xcodebuild -project FHS.xcodeproj -scheme FHS -sdk iphonesimulator
ONLY_ACTIVE_ARCH=NO