Chaning println to print for StoreCreator/main.swift
Motivation:
Swift 2.0 changed the name for println function to print.

I need to investigate if Swift 2.0 allows for a script to import a
module. If that were possible this script could simply import the FHS
module and use it directy, instead of as now being installed and linked
in that process to then be executed which is messy.